Improve dev workflow, use sass in Chrome
Chrome has great support for Sass directly in the browser. You can map
the files served at http://localhost:4000/ to developer-site. Combined
with the source map for the sass files, you can jump directly to the
uncompiled versions of the sass files directly from chrome devtools.
Jekyll does not generate source maps, so this requires a bit of hacking
to set up locally.

    # install sass
    gem install sass
    # make output directory
    mkdir _sass.build
    # use the standalone tool to compile sass
    sass --sourcemap --watch _sass:_sass.build

With this, you can point the link element to _sass.build/main.css and
get all of the nice chrome features.