Remove JSON schema
I tried rendering the JSON schema for booking as a table, but it didn't
exactly read well, so it wasn't very helpful. I wasn't able to come up
with a readable way of presenting the schema, so I'm removing it now
before I waste more time on it.

We don't have schemas today, so we would have to make schemas for all
the APIs, which would be time-consuming. If we find that we need it
later, we must first come up with a nice way to display it.