Add collapsible.js
Collapsible is based on a `toggler` element and a `target` element.

    <a href="#" data-collapse="#target-element" class="collapsible__toggler--collapsed">Toggle an element</a>
    <div id="target-element" class="collapsible__item--collapsed"></div>

The script will read the `data-collapse` attribute and add a click
handler for the toggler element.

The toggler will switch between the following classes for the toggler
and the target element:

    collapsible__toggler--collapsed
    collapsible__toggler--expanded
    collapsible__item--collapsed
    collapsible__item--expanded

That means the actual styling must be done with CSS.