Prevent language variant of getting started page
The page is a category, which in Hugo terms is a collection of the
subpages, but we use it as a standalone page (because of historical
reasons?) which works fine without hacks, but when one of the pages in
that category is translated, the category page also get the same
functionality. The same goes for the home page, but that has an easier
check that was already in place.