Tighten relation between inputs and dl for oneOf
The dl elements corresponding to the selected radio inputs weren’t found
We looked for the closest element type instead of using specific naming
on the elements since we have to select multiple dl elements within a
group that corresponds to the fieldset of the radio inputs.
We had multiple unique names on various attributes within the same
element, the attributes are needed, but we can reuse the unique name.
Separation of JSON and XML also seems to make things clearer.